+++ src/core/model/table/relational_schema.h
@@ -34,7 +34,7 @@ public:
     std::string GetName() const { return name_; }
     std::vector<std::unique_ptr<Column>> const& GetColumns() const { return columns_; };
     Column const* GetColumn(const std::string& col_name) const;
-    Column const* GetColumn(int index) const;
+    Column const* GetColumn(size_t index) const;
     size_t GetNumColumns() const;
     Vertical GetVertical(boost::dynamic_bitset<> indices) const;
